typescript and glimmer change
