+++ packages/core/addon/components/number-format-selector.hbs
@@ -1,4 +1,4 @@
-{{!-- Copyright 2020, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
+{{!-- Copyright 2021, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
 <div class="number-format-selector" ...attributes>
   <DenaliInputGroup @label="Format Type" @isStacked= {{true}} class="number-format-selector__section-header">
     <DenaliRadio
